+++ app/routes/todos.tsx
@@ -6,23 +6,26 @@ import {getUser, getUserId} from "~/session.server";
 import invariant from "tiny-invariant";
 import {formString} from "~/utils";
 import {IsComplete} from "~/routes/todos+/is_complete";
+import {UserRecord} from "~/models/user.server";
 
 export const loader = async (
     {request,}: LoaderFunctionArgs
 ): Promise<TypedResponse<{
     tasks: TaskRecord[];
     q: string;
-    user: never;
+    user: UserRecord;
 }>> => {
     const url = new URL(request.url);
 
     const user = await getUser(request);
     invariant(user?.id, "Missing user");
 
-    const {data, error} = await tasksRepository.getAll(user.id);
-    if (error) {
-        console.error(error);
+    const {data, error} = await tasksRepository.getAll({user_id: user.id});
+    if (error || !data) {
+        console.error("Error fetching task:", error);
+        throw new Response("Task not found", {status: 404});
     }
+
     const q = url.searchParams.get("q") || "";
 
     return json(
@@ -30,7 +33,10 @@ export const loader = async (
     );
 };
 
-export const action = async ({request}: ActionFunctionArgs) => {
+type TodosActionType = { success: boolean }
+export const action = async (
+    {request}: ActionFunctionArgs
+): Promise<TodosActionType> => {
     const userId = await getUserId(request);
 
     const formData = await request.formData();
@@ -42,6 +48,7 @@ export const action = async ({request}: ActionFunctionArgs) => {
 
     if (_action == "create_new") {
         const todolist: TaskMutation = {
+            completed_at: null, deleted_at: null,
             is_complete: false, user_id: userId,
             task: values.task
         };
@@ -49,29 +56,34 @@ export const action = async ({request}: ActionFunctionArgs) => {
         const {data, error} = await tasksRepository.create(todolist);
         console.debug({data, error})
     }
-
     if (_action === "toggle_is_complete") {
-        const {id, is_complete} = values;
-        const isCompleteBool = is_complete === 'true';
-        const {data, error} = await tasksRepository.set(id, {
-            is_complete: isCompleteBool
+        const {id} = values;
+        const todoId = Number(id);
+        const getTaskResponse = await tasksRepository.get({user_id: userId, id: todoId});
+        if (getTaskResponse.error || !getTaskResponse.task) {
+            return {success: false};
+        }
+        const {data, error} = await tasksRepository.set({user_id: userId, id: todoId}, {
+            completed_at: getTaskResponse.task.completed_at ? null : new Date().toISOString(),
         });
-        console.debug({data, error});
+        if(error || !error) {
+            return {success: false};
+        }
+        console.debug({data});
     }
-
     if (_action === "destroy") {
         const {id} = values;
-        const {data, error} = await tasksRepository.destroy(id);
+        const todoId = Number(id);
+        const {data, error} = await tasksRepository.destroy({user_id: userId, id: todoId});
         console.debug({data, error});
     }
-
     return {success: true};
 };
 
 export default function Todos() {
     const {tasks, q, user} = useLoaderData<typeof loader>();
     const navigation = useNavigation();
-    const actionData = useActionData();
+    const actionData = useActionData<TodosActionType>();
     const taskRef = React.useRef<HTMLInputElement>(null);
 
     React.useEffect(() => {
@@ -99,15 +111,15 @@ export default function Todos() {
 
     return (
         <>
-            <div className="flex flex-row w-full h-full p-5">
-                <div className="flex flex-col h-full">
+            <div className="flex flex-row  h-full p-5">
+                <div className="flex flex-col h-full max-w-125">
                     <div className="h-36 flex items-center justify-center">
                         <h1>{user.email}</h1>
                     </div>
                     <div className="flex items-center justify-center pl-1">
                         {/*<h1>My Tasks</h1>*/}
                     </div>
-                    <Form method="post" className="flex w-full justify-between p-2">
+                    <Form method="post" className="flex  justify-between p-2">
                         <input type="text" name="task" id="taskTitle"
                                placeholder="Enter task title"
                                ref={taskRef}
@@ -136,12 +148,12 @@ export default function Todos() {
                             hidden={!searching}
                         />
                     </Form>
-                    <nav className="flex w-full justify-center p-2">
+                    <nav className="flex  justify-center p-2">
                         {tasks && tasks.length ? (
-                            <ul className=" w-full">
+                            <ul className=" ">
                                 {tasks.map((task) => (
                                     <li key={task.id}
-                                        className="w-full flex flex-row justify-between">
+                                        className=" flex flex-row justify-between">
                                         <IsComplete task={task}/>
 
                                         <NavLink
@@ -151,10 +163,13 @@ export default function Todos() {
                                             }
                                             to={`${task.id}`}
                                         >
-                                            <div className="flex cursor-pointer items-center hover:bg-gray-300 ">
-
-                                                <p>{task.task}</p>
-                                            </div>
+                                            <h3 className="
+                                            overflow-hidden truncate flex cursor-pointer items-center
+                                            font-medium
+                                            hover:bg-red hover:text-white
+                                            ">
+                                                {task.task}
+                                            </h3>
                                         </NavLink>
 
                                         <Form
@@ -186,10 +201,8 @@ export default function Todos() {
                                                     <path strokeLinecap="round" strokeLinejoin="round"
                                                           d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0"/>
                                                 </svg>
-
                                             </button>
                                         </Form>
-
                                     </li>
                                 ))}
                             </ul>
